patch version

+++ includes/FDiskData.h
@@ -6,7 +6,7 @@ struct FDiskData
 {
 	timestamp 	*recordingDateAndTime;
 	dstring		volumeIdentifier[32];
-	uint16_t	version;
+	double		version;
 	uint64_t	totalSpace;
 	uint64_t	freeSpace;
 };
